refactored and created a wiki helpfer functions class to reduce complexity
